==> R/notebooks/3_plots_gage_discharge_jrc_dswe2_by_CV_huc.Rmd <==
---
title: "Stream gage correlations with JRC and DSWE data in California's Central Valley - Result and plots"
fontsize: 9
output: html_notebook
editor_options:
  chunk_output_type: inline
---


```{r, echo=FALSE, warning=FALSE, message=FALSE}

## ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
## 3_plots_gage_discharge_pekel_dswe2_correlations_by_CV_huc.Rmd
## ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# ---------------------------------------------------------------------- #
#### General setup ####
# ---------------------------------------------------------------------- #

# Function: Load packages
loadPackages <- function(pkgs){
  for (pkg in pkgs){
    if(!require(pkg, character.only = TRUE)){
      install.packages(pkg, dependencies = TRUE)
      require(pkg, character.only = TRUE)
    }
  }
}

# Load packages
loadPackages(c("dataRetrieval", # 2.7.4
               "plyr",          # 1.8.4
               "tidyverse",     # 1.2.1
               'data.table',    # 1.12.0
               'knitr',         # 1.21
               'kableExtra',    # 1.0.1
               'devtools',      # 2.0.1
               'googledrive',   # 0.1.3
               'zoo'))          # 1.8-4

# Load mapping packages
loadPackages(c("ggmap",         # 3.0.0
               "sf",            # 0.7-3
               "maps",          # 3.3.0
               "mapdata",       # 2.3.0
               'rgdal',         # 1.4-4
               'plotly'))       # 4.8.0

# Set default as echo FALSE
knitr::opts_chunk$set(echo = FALSE)

```


```{r, echo = FALSE}

# Plotly setup
Sys.setenv("plotly_username"= "jjwalker")
Sys.setenv("plotly_api_key"="KXnTHoAperT2jUBO89Xr")
```


```{r, echo=FALSE}

# ---------------------------------------------------------------------- #
#### Set correlation threshold ####
# ---------------------------------------------------------------------- #

# Identifies threshold between good and bad correlations

cor_threshold <- 0.7

```


```{r, echo=FALSE}

# ---------------------------------------------------------------------- #
#### Set NoData threshold and correlation type ####
# ---------------------------------------------------------------------- #

# These are not dynamic; they identify which file to load

nodata_threshold <- 5

# Correlation test used for gage to imagery comparisons

cor_type <- 'spearman' #'pearson' # 'kendall'

```


```{r, echo = FALSE, include = F, results = 'hide'}

# ---------------------------------------------------------------------- #
#### Set files and folders ####
# ---------------------------------------------------------------------- #

# Plots are saved to path.out/plots

path.data <-  "E:/projects/place"
file.in <- file.path(path.data, "data", paste0("CV_dswe_jrc_gage_", nodata_threshold, "pct_nodata_correlations_", cor_type,".RData"))

# Set path to shapefile directory
shp.dir <- file.path(path.data, "R", "shp")
ifelse(!dir.exists(shp.dir), dir.create(shp.dir), FALSE)

```


```{r, echo=FALSE}

# ---------------------------------------------------------------------- #
#### Load file of processed data ####
# ---------------------------------------------------------------------- #

load(file.in)

```
 

```{r, echo = FALSE, message = FALSE, include = FALSE}

# ---------------------------------------------------------------------- #
#### Download HUC GIS data from Google Drive ####
# ---------------------------------------------------------------------- #

# Download the shapefile if it doesn't exist
  shp.name <- 'WBDHU8_Central_Valley_UTMz10.shp'

if (!file.exists(file.path(shp.dir, shp.name))) {
  data_url = "1_2JnwZFS0blok23-9O7lw-o7Hk96E-nc" 
  download.file(sprintf("https://docs.google.com/uc?id=%s&export=download", data_url), 
                destfile = file.path(shp.dir, "huc.zip"), mode = 'wb')
  unzip(zipfile = file.path(shp.dir, 'huc.zip'), exdir = shp.dir)
  file.remove(file.path(shp.dir, 'huc.zip'))
}

shp <- st_read(dsn = file.path(shp.dir, shp.name), stringsAsFactors = F)

```


####Workflow

* Load R image file of correlations between stream gage info and JRC/DSWE data
* Calculate and/or plot relevant statistics
  - Discharge/imagery correlation type: **`r toupper(cor_type)`** 
  - Clear-pixel threshold value: **`r nodata_threshold`**%


####Inputs

  - R image file generated from **process_gage_discharge_pekel_dswe2_correlations_by_CV_huc.Rmd** stored locally on `r file.path(path.data, "data")`
  - Shapefile of Central Valley HUCs stored on ScienceBase (Google Drive placeholder for now)
  

####Outputs

  - In-line results
  - Plots of augmented JRC and DSWE time series based on within- and external-HUC gage correlations to **`r file.path(path.data, "plots")`**
  
<br>  


####Central Valley HUCs
  
```{r, echo = FALSE, message = FALSE}

# ---------------------------------------------------------------------- #
#### Plot HUCs shapefile ####
# ---------------------------------------------------------------------- #

shp %>% 
ggplot() + 
  geom_sf(aes(fill = HUC8), show.legend = FALSE) +  # FALSE for no legend
  theme_bw()  +
  coord_sf(crs = "+init=epsg:4269")  #4269
  coord_sf(crs = st_crs(26910)) # set UTM zone 10 projection

# http://eriqande.github.io/rep-res-web/lectures/making-maps-with-R.html



```
  

  
####NoData proportions

JRC and DSWE data show different proportions of NoData pixels (clouds, cloud shadows, or snow). In general, DSWE data have lower proportions of contaminated pixels. A seasonal breakdown shows that DSWE winter and fall images tend to have more "workable" data than JRC images. The quality of that data is unknown, however. 

```{r, echo = F, warning = FALSE}

# ---------------------------------------------------------------------- #
#### Plot boxplot of NoData proportions ####
# ---------------------------------------------------------------------- #

pk_ds_season <- pk_ds
yq <- as.yearqtr(as.yearmon(pk_ds_season$date, "%Y-%m-%d") + 1/12)
pk_ds_season$season <- factor(format(yq, "%q"), levels = 1:4, 
labels = c("Winter", "Spring", "Summer", "Fall"))

# Figure for display
ggplot(pk_ds_season, aes(season, pct, color = type)) + 
  geom_boxplot() +
  theme_bw() +
  theme(axis.text = element_text(size = 10),  #12
        axis.title = element_text(size = 12), #14
        axis.text.x = element_text(vjust = 0.2),
        axis.title.x = element_blank(), 
        legend.text = element_text(size = 10),  #12
        legend.title = element_text(size = 12)) +  #14
  ylab("% NoData pixels") +
  scale_color_discrete(name = "Dataset", breaks = c("dswe", "pekel"), labels = c("DSWE", "JRC"))


# Figure for publication
# ggplot(pk_ds_season, aes(season, pct, color = type)) + 
#   geom_boxplot(lwd = 0.46, outlier.size = 0.4) +
#   theme_bw(base_size = 5) +
#   theme(axis.text = element_text(size = 8),  
#         axis.title = element_text(size = 10), 
#         axis.text.x = element_text(vjust = 0.2),
#         axis.title.x = element_blank(), 
#         legend.text = element_text(size = 8),  
#         legend.title = element_text(size = 10)) +  
#   ylab("% NoData pixels") +
#   scale_color_discrete(name = "Dataset", breaks = c("dswe", "pekel"), labels = c("DSWE", "JRC"))
# 
# ggsave(file = file.path(path.data, "plots", paste0("nodata_pixels.pdf")),
#         width = 4.5, #7
#         height = 2.8, #4
#         units = "in",
#         dpi = 600)

```
####General gage/discharge stats

<br>
Total number of gages across the CV

```{r, echo = FALSE}

n_gages <- pk_ds_sg.int %>%
  ungroup() %>% 
  summarize(n_all = n_distinct(site_no)) 

n_gages <- n_gages[[1]]

n_gages %>% 
   kable(col.names = c("n")) %>%  
     kable_styling("striped", full_width = FALSE, position = 'left') 
```

<br>
Range of monthly mean gage discharge (cu ft/s)
```{r, echo = FALSE}

pk_ds_sg.int %>%
  group_by(site_no, huc8.img, type, cor) %>% 
  summarise(mean = mean(mean_va)) %>% 
  ungroup() %>% 
  summarise(min = min(mean), minSite = site_no[which.min(mean)],
            max = max(mean), maxSite = site_no[which.max(mean)]) %>% 
  kable(col.names = c("Min", "Gage #", 'Max', 'Gage #'), digits = 3) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left') 

```
<br>
All within-HUC correlations above/below `r cor_threshold` 

```{r, echo = FALSE}


# --------------------------------------------------------
# Number of within-HUC correlations above threshold ####
# --------------------------------------------------------

pk_ds_sg.int %>%
  filter(huc8.img == huc8.sg) %>% 
  group_by(type) %>% 
  distinct(huc8.img, cor) %>%
  filter(cor > cor_threshold) %>% 
  summarise(n_over = n(),
            pct_gt = n_over/n_gages * 100,
            n_under = n_gages - n(),
            pct_lt = n_under/n_gages * 100) %>% 
    
   kable(col.names = c("Dataset", "# greater than ", '%', "# less than", "%"), digits = 1) %>%  
     kable_styling("striped", full_width = FALSE, position = 'left') 
            

```
<br>
Within-HUC correlations greater than `r cor_threshold` - max r in each HUC

```{r, echo = FALSE}

# *** hard-coded value for the number of HUCs

cor.huc.max %>% 
  group_by(type) %>%
  filter(cor > cor_threshold) %>% 
  summarize(n = n(),
            pct = n/49*100) %>% 

 kable(col.names = c("Type", "n", "% of HUCs"), digits = 1) %>%  
     kable_styling("striped", full_width = FALSE, position = 'left')  
```

Range of maximum within-HUC correlations

```{r, echo = FALSE}
# ---------------------------------------------------------------------- #
# Correlation range for MAX corr - JRC and DSWE vs. stream gage ####
# ---------------------------------------------------------------------- #

cor.huc.max %>%
  group_by(type) %>% 
  summarise(min_cor = min(abs(cor)),
            max_cor = max(cor)) %>% 
  kable(col.names = c("Dataset", "Min r", "Max r"), digits = 3) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left')



```

####Maximum within-HUC correlation in each HUC
```{r, echo = FALSE, warning=FALSE,  fig.width=8, fig.height=6}

# ---------------------------------------------------------------------- #
# Correlations for JRC and DSWE vs. stream gage - PLOT ####
# ---------------------------------------------------------------------- #

p <- ggplot(cor.huc.max, aes(x = reorder(huc8.img, -cor), y = cor, fill = type)) + 
       geom_bar(stat = 'identity', position = position_dodge()) +
       geom_abline(slope = 0, intercept = cor_threshold, col = "black", lty = 2) +
       theme_bw() +
       theme(axis.text.x = element_text(angle = 80, hjust = 1),
             legend.title = element_blank()) +
       xlab('HUC') +
       ylab('Correlation') +
       scale_fill_manual(name = "Dataset", 
                    values = c('#F8766D', '#00BFC4'), 
                    breaks = c("dswe", "pekel"),
                    labels = c("DSWE", "JRC"))  # plotly doesn't honor this 

# Plot for display
ggplotly(p, tooltip = c('huc8', 'cor')) %>% 
  add_annotations(text="Dataset", xref="paper", yref="paper",
                  x=1.02, xanchor="left",
                  y=0.8, yanchor="bottom",   
                  legendtitle = TRUE, 
                  showarrow = FALSE) %>%
  layout(legend = list(y = 0.8, yanchor = "top"))

# Plot for publication
 #  ggplot(cor.huc.max, aes(x = reorder(huc8.img, -cor), y = cor, fill = type)) + 
 #       geom_bar(stat = 'identity', position = position_dodge(), width = 0.75) +
 #       geom_abline(slope = 0, intercept = cor_threshold, col = "black", lty = 2, size = 0.4) +
 #       theme_bw(base_size = 8) +
 #       theme(axis.text.x = element_text(angle = 65, hjust = 1, size = 7),
 #             axis.text.y = element_text(size = 7),
 #             legend.title = element_blank(),
 #             legend.key.size = unit(0.8, 'line'),
 #             legend.text = element_text(size = 7)) +
 #       xlab('HUC') +
 #       ylab('Correlation') +
 #       scale_fill_manual(name = "Dataset", 
 #                    values = c('#F8766D', '#00BFC4'), 
 #                 #   values = c('orange', 'lightblue'),
 #                    breaks = c("dswe", "pekel"),
 #                    labels = c("DSWE", "JRC"))
 # 
 # ggsave(file = file.path(path.data, "plots", paste0("max correlation_withinHUC_", cor_type, ".pdf")),
 #         width = 6.4, #7
 #         height = 3.2, #4
 #         units = "in",
 #         dpi = 600)


```

```{r, echo=FALSE, warning=FALSE, message = FALSE, include = FALSE}
####Maximum within-HUC correlation in each HUC - TABLE
# ---------------------------------------------------------------------- #
# Display correlation table for JRC and DSWE vs stream gage = TABLE ####
# ---------------------------------------------------------------------- # 

 cor.huc.max %>%
     select(huc8.img, site_no, parameter_cd, type, cor, n) %>% 
     kable(col.names = c("HUC", "Gage", "Gage type", "Data type", "r", "months of data"), digits = 3) %>%  
     kable_styling("striped", full_width = FALSE) %>%
     row_spec(c(which(cor.huc.max$cor > cor_threshold)), bold = T) %>%    
     scroll_box(height = '400px', width = '500px')
  
```


<br>

####Within-HUC correlation range and # of gages in each HUC
```{r, echo = FALSE}
# ---------------------------------------------------------------------- #
# Correlation range for individual HUCSs ####
# ---------------------------------------------------------------------- #

pk_ds_sg.int %>%
  filter(huc8.img == huc8.sg) %>% 
  distinct(cor, huc8.img, type)  %>% 
  group_by(huc8.img, type) %>% 
  summarise(min_cor = min(abs(cor)),
            max_cor = max(cor),
            n = n()) %>% 
  kable(col.names = c("HUC", "Dataset", "Min", "Max", "# gages"), digits = 3) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left') %>% 
  scroll_box(height = '400px', width = '400px')


```
<br>

####Correlations in HUCs w different numbers of gages
```{r, echo = FALSE,  warning = FALSE, message=FALSE}
# ---------------------------------------------------------------------- #
# Mean of the maximums correlations in each HUC, by group ####
# ---------------------------------------------------------------------- #

gage_thresh <- 6

temp <- pk_ds_sg.int %>%
  filter(huc8.img == huc8.sg) %>%  # each gage will have data for either dataset
  distinct(huc8.img)  %>% 
  group_by(huc8.img, type) %>% 
  summarise(n_gages = n(),
            max_cor = max(cor)) %>%
  mutate(cat = if(n_gages < gage_thresh) paste0("< ", gage_thresh) else paste0(">= ", gage_thresh)) 

temp %>% 
  group_by(type, cat) %>% 
  summarise(n_total = n(),
            cor = mean(max_cor)) %>%   
  kable(col.names = c("Dataset", "Gages/HUC",  "# HUCs", 'Mean max cor'), digits = 3) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left')

```

<br>

#### # of gages in # of HUCs
```{r, echo = FALSE}
# ---------------------------------------------------------------------- #
# Number of gages in each HUC ####
# ---------------------------------------------------------------------- #

pk_ds_sg.int %>%
  ungroup() %>% # added bc slooow otherwise
  filter(huc8.img == huc8.sg & type == "dswe") %>%  # each gage will have data for either dataset
  group_by(huc8.img) %>% 
  summarize(n_gages = n_distinct(site_no)) %>% 
  ungroup() %>% 
  group_by(n_gages) %>% 
  summarize(n_hucs = n()) %>% 
  kable(col.names = c("# of gages", "# HUCs"), digits = 3) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left') %>% 
  scroll_box(height = '400px', width = '400px')

```
<br>

####Mean and median # of gages per HUC

```{r, echo = FALSE}
# ---------------------------------------------------------------------- #
# Mean and median # of gages/HUC ####
# ---------------------------------------------------------------------- #
# Need to incorporate the missing HUC(s) to ensure the mean is calculated correctly.
# The missing HUC is currently hard-coded in.  Could get it programmatically by:

# setdiff(unique(pk_ds_sg.int$huc8.img), unique(pk_ds_sg.int$huc8.sg))

mean_n_gages <- pk_ds_sg.int %>%
  ungroup() %>% 
  filter(huc8.img == huc8.sg & type == 'dswe') %>% 
  distinct(site_no, huc8.img)  %>% 
  group_by(huc8.img) %>% 
  summarise(n = sum(n())) 

missing_huc <- c('18030004', 0)

# Add missing gage to
mean_n_gages <- rbind(missing_huc, mean_n_gages)
mean_n_gages <- as.data.frame(mean_n_gages)


mean_n_gages %>% 
  ungroup() %>% 
  mutate(n = as.numeric(n)) %>% 
  summarise(mean = mean(n),
            med = median(n)) %>% 
  kable(col.names = c("mean # gages/HUC", "median # gages/HUC"), digits = 3) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left')


```
<br>

####Long-term time series

```{r, echo = FALSE, warning = FALSE}

n_years <- 30

pk_ds_sg.series <-   
  pk_ds_sg.int %>%
  group_by(site_no, huc8.img, huc8.sg, type, cor) %>% 
  summarize(n_months = n()) %>% 
  filter(n_months > 12 * n_years & cor > 0.7) %>% 
  group_by(huc8.img) %>% 
  slice(which.max(cor))

pk_ds_sg.int.series <- 
  inner_join(pk_ds_sg.series, pk_ds_sg.int, by = c("site_no", "huc8.img", "huc8.sg", "type", "cor"))

#write.csv(pk_ds_sg.int.series, file = "E:/projects/place/output/tables/pk_ds_sg_gt30years_and_corgt0pt7.csv", row.names = FALSE)

```

Internal-match HUCs with 30+ years of data and a correlation basis > 0.7. _Max correlations only_
```{r, echo = FALSE}

pk_ds_sg.series %>% 
  group_by(site_no, huc8.img, huc8.sg, type, cor) %>% 
  filter(huc8.img == huc8.sg) %>% 
  group_by(huc8.img) %>% 
  slice(which.max(cor)) %>% 
  distinct(site_no, huc8.img, huc8.sg, type, cor) %>% 
  arrange(huc8.img, site_no) %>% 
  kable(col.names = c("Gage", "Imagery HUC", "Gage HUC", "Dataset", 'Corr'), digits = 2) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left') 


```

<br>

External-match HUCs with 30+ years of data and a correlation basis > 0.7. _Max correlations only_
```{r, echo = FALSE}

pk_ds_sg.series %>% 
  group_by(site_no, huc8.img, huc8.sg, type, cor) %>% 
  filter(huc8.img != huc8.sg) %>% 
  group_by(huc8.img) %>% 
  slice(which.max(cor)) %>% 
  distinct(site_no, huc8.img, huc8.sg, type, cor) %>% 
  arrange(huc8.img, site_no) %>% 
  kable(col.names = c("Gage", "Imagery HUC", "Gage HUC", "Dataset", 'Corr'), digits = 2) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left') %>% 
    scroll_box(height = '400px', width = '450px')


```

<br>

Number of HUCs with 30+ years of data and a correlation basis > 0.7
```{r, echo = FALSE}

pk_ds_sg.int.series %>% 
  ungroup() %>% 
  distinct(huc8.img) %>% 
  summarize(n = n()) %>% 
  kable(col.names = c("# of HUCs")) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left')


```

####Correlation between COV and r values

```{r, echo = FALSE}
corr_stats <-  pk_ds_sg.int %>%
  group_by(site_no, huc8.img, type, cor) %>% 
  summarise(sd = sd(mean_va),
            mean = mean(mean_va),
            cov = sd/mean) 

cor(corr_stats$cor, corr_stats$cov) %>% 
  kable(col.names = c("r"), , digits = 3) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left')


```


<br>

####DSWE vs JRC within-HUC correlations

JRC correlations are stronger in almost all HUCs. HUC 18030004 is the only HUC in which no gage matched the requirement of having at least 5 years of overlapping data with the imagery time series. 


```{r, echo = FALSE}

# ---------------------------------------------------------------------- #
#### Table of whether DSWE or JRC has the max correlation ####
# ---------------------------------------------------------------------- #

  cor.huc.max %>%
    group_by(huc8.img) %>%
    slice(which.max(cor)) %>% 
    group_by(type) %>% 
    count(type) %>% 
   kable(col.names = c("Dataset", "# max corr"), digits = 0) %>%  
   kable_styling("striped", full_width = FALSE, position = 'left') 
  
  
```




```{r, echo = FALSE}

# ---------------------------------------------------------------------- #
#### Spatial plot of whether DSWE or JRC has higher discharge correlation ####
# ---------------------------------------------------------------------- #

maxes.by.type <- cor.huc.max %>%
    group_by(huc8.img) %>%
    slice(which.max(cor))
 
shp %>%
   mutate(HUC8 = as.numeric(HUC8)) %>% 
   left_join(maxes.by.type, c("HUC8" = "huc8.img")) %>% 
   ggplot() + 
     geom_sf(aes(fill = type), size = 0.1) +
     theme_bw(base_size = 10) +
     theme(plot.margin=grid::unit(c(1, 0, 0, 0), "mm"),  #(T, R, B, L)
           legend.text = element_text(size = 8),
           legend.key = element_rect(size = 1, fill = "white", color = "white"),
           legend.key.size = unit(0.8, 'lines'),
           legend.spacing.y = unit(1, 'mm')) +  
     scale_size_identity() +
     scale_y_continuous(breaks = c(35, 37, 39, 41)) +
     scale_x_continuous(breaks = c(-122, -120, -118)) +
     scale_fill_manual(name = "Imagery \ndataset", 
                         #values= c('orange', 'lightblue'), 
                         values = c('#F8766D', '#00BFC4'), 
                         labels = c('DSWE', "JRC")) +
 #    coord_sf(crs = st_crs(26910)) # set USA_Contiguous_Albers_Equal_Area_Conic projection
     coord_sf(crs = "+init=epsg:4269")  #NAD83 GEO   

# Save for publication
 ggsave(file = file.path(path.data, "plots", paste0("imagery_w_higher_cor_", cor_type, ".pdf")),
         width = 3.2,
         height = 3.6,
         units = "in",
         dpi = 600)



```
<br>

####Months of data available for each best within-HUC correlation
```{r, echo = FALSE, warning = FALSE, message = FALSE }


# ---------------------------------------------------------------------- #
#### Spatial plot of # months available for each 'best' correlation ####
# ---------------------------------------------------------------------- #

shp %>%
   mutate(HUC8 = as.numeric(HUC8)) %>% 
     inner_join(maxes.by.type, c("HUC8" = "huc8.img")) %>% 
  ggplot() + 
     theme_nothing() +
     geom_sf(aes(fill = n)) +
     theme_bw() +
    scale_fill_gradient(name = "Months of data", limits = c(0, max(cor.huc.max$n)), low = 'lightgreen', high = 'blue') +
     coord_sf(crs = st_crs(4269)) # set UTM zone 10 projection 26910
```
<br>

#### EXTERNAL GAGES

####Gage:imagery correlations across all HUCs  

Opening up the analysis to all gages, no matter what HUC they're in, reveals that the majority of imagery datasets have a stronger correlation with a gage that is outside that particular HUC. 

```{r, echo = FALSE, warning=FALSE,  fig.width=8, fig.height=6}

# ---------------------------------------------------------------------- #
#### Bar plot of max HUC correlations -- All gages ####
# ---------------------------------------------------------------------- #

p <- ggplot(cor.all.max, aes(x = reorder(huc8.img, -cor), y = cor, fill = type)) + 
       geom_bar(stat = 'identity', position = position_dodge()) +
       geom_abline(slope = 0, intercept = cor_threshold, col = "black", lty = 2) +
       theme_bw() +
       theme(axis.text.x = element_text(angle = 80, hjust = 1),
             legend.title = element_blank()) +
       xlab('HUC') +
       ylab('Correlation') +
       scale_fill_manual(name = "Dataset", 
                    values = c('#F8766D', '#00BFC4'), 
                #   values = c('orange', 'lightblue'),
                    breaks = c("dswe", "pekel"),
                    labels = c("DSWE", "JRC"))


ggplotly(p, tooltip = c('huc8', 'cor')) %>% 
  add_annotations(text = "Dataset", xref = "paper", yref = "paper",
                  x=1.02, xanchor="left",
                  y=0.8, yanchor="bottom",   
                  legendtitle = TRUE, 
                  showarrow = FALSE) %>%
  layout(legend = list(y = 0.8, yanchor = "top"))


# Plot for publication
 #  ggplot(cor.all.max, aes(x = reorder(huc8.img, -cor), y = cor, fill = type)) +
 #       geom_bar(stat = 'identity', position = position_dodge(), width = 0.75) +
 #       geom_abline(slope = 0, intercept = cor_threshold, col = "black", lty = 2, size = 0.4) +
 #       theme_bw(base_size = 8) +
 #       theme(axis.text.x = element_text(angle = 65, hjust = 1, size = 7),
 #             axis.text.y = element_text(size = 7),
 #             legend.title = element_blank(),
 #             legend.key.size = unit(0.8, 'line'),
 #             legend.text = element_text(size = 7)) +
 #       xlab('HUC') +
 #       ylab('Correlation') +
 #       scale_fill_manual(name = "Dataset",
 #                    values = c('#F8766D', '#00BFC4'),
 #                 #   values = c('orange', 'lightblue'),
 #                    breaks = c("dswe", "pekel"),
 #                    labels = c("DSWE", "JRC"))
 # 
 # ggsave(file = file.path(path.data, "plots", paste0("max correlation_allHUC_", cor_type, ".pdf")),
 #         width = 6.4, #7
 #         height = 3.2, #4
 #         units = "in",
 #         dpi = 600)

```

<br>

Red = external-HUC match; bold > 0.7

```{r, echo = FALSE, warning = FALSE} 

# ---------------------------------------------------------------------- #
#### Display gage with max correlation in each HUC - ALL GAGES ####
# ---------------------------------------------------------------------- #

# cor.all.max has the maximum within-huc correlation and the maximum outside-huc correlation for each HUC

tmp1 <- cor.all.max %>% 
  group_by(huc8.img, type) %>%
  slice(which.max(cor)) %>% 
  select(huc8.img, site_no, huc8.sg, type, cor, same_huc, n_site)

# Breaking it up like this is necessary b/c kable can't handle the fact that cor.all.max has changed ('select' messes it up)
tmp1 %>% 
  kable(col.names = c("HUC - imagery", "Gage w best correlation", "HUC - gage", "Data type", "r", 'Same HUC?','times gage is best'), digits = 3) %>%
  kable_styling("striped", full_width = FALSE, position = 'left') %>%
  row_spec(c(which(tmp1$cor > cor_threshold)), bold = T) %>%
  row_spec(c(which(tmp1$same_huc == 'N')), color = "red") %>% 
  scroll_box(width = '600px', height = '400px')

```
<br>
Number of gages that correspond best to imagery within/external to same HUC
```{r, echo = FALSE, warning = FALSE} 

# ---------------------------------------------------------------------- #
#### Quantify number of gages that perform better - within-HUC/all gages
# ---------------------------------------------------------------------- #

# cor.all.max has the maximum within-huc correlation and the maximum outside-huc correlation for each HUC

tmp1 %>% 
  ungroup() %>% 
  group_by(type, same_huc) %>% 
  summarize(n = n()) %>% 
 kable(col.names = c("Dataset", "Same HUC", "n"), digits = 1) %>%  
     kable_styling("striped", full_width = FALSE, position = 'left')  


```
<br>

Range of maximum all-HUC correlations

```{r, echo = FALSE}
# ---------------------------------------------------------------------- #
# Correlation range for MAX corr - JRC and DSWE vs. stream gage  - ALL-HUC ####
# ---------------------------------------------------------------------- #

cor.all.max %>%
  group_by(type) %>% 
  summarise(min_cor = min(abs(cor)),
            max_cor = max(cor)) %>% 
  kable(col.names = c("Dataset", "Min r", "Max r"), digits = 3) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left')



```
<br>

####All-HUC correlations greater than `r cor_threshold`

```{r, echo = FALSE}

cor.all.max %>% 
  group_by(type) %>%
  filter(cor > cor_threshold) %>% 
  summarize(n = n(),
            pct = n/50*100) %>%  # 50 HUCs b/c the huc without a gage can have a correlation now

 kable(col.names = c("Type", "n", "% of HUCs"), digits = 2) %>%  
     kable_styling("striped", full_width = FALSE, position = 'left')  
```
<br>

####Super gage 

The 'super gage' here is site `r as.character(unlist(cor.all.max[which.max(cor.all.max$n_site), 'site_no']))` in HUC `r as.character(cor.all.max[which.max(cor.all.max$n_site), 'huc8.sg'])`, which produces the highest correlations with `r as.character(cor.all.max[which.max(cor.all.max$n_site), 'n_site'])` HUCs
```{r, echo = FALSE}

# ---------------------------------------------------------------------- #
#### Spatial plot of super gage correlation -- DSWE ####
# ---------------------------------------------------------------------- #

# Get the correlation data into the shapefile
shp_cor <- merge(shp, cor.all.max, by.x = 'HUC8', by.y = 'huc8.sg')
n_max <- max(shp_cor$n_site)

# Get the site coordinates
site_coords <- cor.all.max %>% 
               filter(type == 'pekel') %>% 
               select(site_no, lat, lon, n_site)
site_coords <- site_coords[complete.cases(site_coords), ]

# Transform shp polygons to match WGS84 lat/lon coords
shp_4326 <- st_transform(shp, '+init=epsg:4326')
shp_cor_4326 <- st_transform(shp_cor, '+init=epsg:4326')

# Get hucs that use the super gage
super_hucs <- unique(as.data.frame(shp_cor[shp_cor$n_site == n_max, ])['huc8.img'])
super_hucs$huc8.img <- as.character(super_hucs$huc8.img)

# Plot super gage and associated HUCs
shp_4326 %>% 
     right_join(super_hucs, by = c("HUC8" = 'huc8.img')) %>% 
  ggplot(.) +  # plot the joined hucs (subset)
  geom_sf(fill = 'red', color = 'gray50', lwd = 0.4) +  # associated hucs only
     geom_sf(data = shp_4326, fill = NA, color = 'gray50', lwd = 0.4) +  # all hucs
     geom_point(data = subset(site_coords, n_site == n_max), aes(x = lon, y = lat), color = "black", shape = 18, size = 6) + 
  theme_bw()


```

<br>


####Max correlations with DSWE class combinations

Gage discharge and "highest confidence" water extent generally have the highest correlations. The list contains only the top class 1 correlation in each HUC.

```{r, echo = FALSE}
 
tmp2 <- ds_sg.cor %>% 
      ungroup(huc8.sg) %>% 
      select(huc8.img, site_no, cor1, cor12, cor123, cor1234, n) %>% 
      group_by(huc8.img) %>% 
      slice(which.max(cor1))

tmp2 %>% 
      kable(col.names = c("HUC", "Site", "r (w1)", "r (w1+w2)", "r (w1+w2+w3)", "r (w1+w2+w3+w4)", "n"), digits = 3) %>%  
      kable_styling("striped", full_width = FALSE, position = 'left') %>%
      row_spec(c(which(tmp2$cor1 > tmp2$cor12)), bold = T) %>% 
      scroll_box(height = '400px', width = '600px')

```

<br>

####Summary of DSWE water1/water1+2 

```{r, echo = FALSE}

ds_sg.cor %>% 
      ungroup() %>% 
      filter(cor1 > cor12) %>% 
      summarize(n_sites = n_distinct(site_no),
            rmse_cor1 = sqrt(mean(cor1 - cor12)^2)) %>% 
  kable(col.names = c("# gages where Water > Water1+2", "RMSE"), digits = 3) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left') 
```
<br>

```{r, echo = FALSE}

ds_sg.cor %>% 
      ungroup() %>% 
      filter(cor12 > cor1) %>% 
      summarize(n = n_distinct(site_no),
            rmse_cor1 = sqrt(mean(cor1 - cor12)^2)) %>% 
  kable(col.names = c("# gages where Water1+2 > Water1", "RMSE"), digits = 3) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left') 

```
<br>

####JRC vs. DSWE imagery correlations 

#####Number of correlations greater than `r cor_threshold`

```{r, echo = FALSE}

# ---------------------------------------------------------------------- #
#### JRC/DSWE correlations ####
# ---------------------------------------------------------------------- #

 
pk_ds.cor %>% 
  filter(cor > cor_threshold) %>% 
  summarize(n = n()) %>% 
  kable(col.names = c("# gages")) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left') 




```
#####Correlations less than `r cor_threshold`
```{r, echo = FALSE}

# ---------------------------------------------------------------------- #
#### JRC/DSWE correlations ####
# ---------------------------------------------------------------------- #
 
pk_ds.cor %>% 
  filter(cor < cor_threshold) %>% 
  kable(col.names = c("HUC", "cor", "# months JRC", "# months DSWE"), digits = 2) %>%  
  kable_styling("striped", full_width = FALSE, position = 'left') 
```


####Correlations - JRC vs. DSWE plot

```{r, echo = FALSE}

# ---------------------------------------------------------------------- #
#### Spatial plot of JRC/DSWE correlations ####
# ---------------------------------------------------------------------- #

 shp %>%
   mutate(HUC8 = as.numeric(HUC8)) %>% 
   left_join(pk_ds.cor, c("HUC8" = "huc8.img")) %>% 
   ggplot() + 
     geom_sf(aes(fill = cor)) +
     theme_bw() +
     scale_fill_gradient(name = "Correlation", limits = c(0,1), low = 'green', high = 'blue') +
     coord_sf(crs = st_crs(26910)) # set UTM zone 10 projection



##### Publication plot
shp %>%
   mutate(HUC8 = as.numeric(HUC8)) %>% 
   left_join(pk_ds.cor, c("HUC8" = "huc8.img")) %>% 
   ggplot() + 
     geom_sf(aes(fill = cor), size = 0.1) +
     theme_bw(base_size = 10) +
     theme(plot.margin=grid::unit(c(2, 0, 0, 0), "mm"),  #(T, R, B, L)
           legend.text = element_text(size = 8),
           legend.key = element_rect(size = 1, fill = "white", color = "white"),
           legend.key.size = unit(1.1, 'lines'),
           legend.spacing.y = unit(1, 'mm')) +  
     scale_size_identity() +
     scale_y_continuous(breaks = c(35, 37, 39, 41)) +
     scale_x_continuous(breaks = c(-122, -120, -118)) +
     scale_fill_manual(name = "Imagery \ndataset", 
                         #values= c('orange', 'lightblue'), 
                         values = c('#F8766D', '#00BFC4'), 
                         labels = c('DSWE', "JRC")) +
     scale_fill_gradient(name = "Correlation", limits = c(0,1), low = 'green', high = 'blue') +
   #  coord_sf(crs = st_crs(26910)) # set UTM zone 10 projection
     coord_sf(crs = "+init=epsg:4269")  #NAD83 GEO   


# Save for publication
 ggsave(file = file.path(path.data, "plots", paste0("dswe_jrc_imagery_cor.pdf")),
         width = 3.5,
         height = 3.8,
         units = "in",
         dpi = 600)



```
<br>


####Correlations - DSWE vs. gage
```{r, echo = FALSE}
# ---------------------------------------------------------------------- #
#### Spatial plot of DSWE/gage correlations ####
# ---------------------------------------------------------------------- #

# Get the max for DSWE by HUC

 shp %>%
   mutate(HUC8 = as.numeric(HUC8)) %>% 
   inner_join(cor.huc.max[cor.huc.max$type == 'dswe', ], c("HUC8" = "huc8.img")) %>% 
   ggplot() + 
     geom_sf(aes(fill = cor)) +
     theme_bw() +
     scale_fill_gradient(limits = c(0,1), low = 'green', high = 'blue') +
     coord_sf(crs = st_crs(26910)) # set UTM zone 10 projection

```
<br>


####Correlations - JRC vs. gage
```{r, echo = FALSE}

# ---------------------------------------------------------------------- #
#### Spatial plot of JRC/gage correlations ####
# ---------------------------------------------------------------------- #


 shp %>%
   mutate(HUC8 = as.numeric(HUC8)) %>% 
   inner_join(cor.huc.max[cor.huc.max$type == 'pekel', ], c("HUC8" = "huc8.img")) %>% 
   ggplot() + 
     geom_sf(aes(fill = cor)) +
     theme_bw() +
     scale_fill_gradient(limits = c(0,1), low = 'green', high = 'blue') +
     coord_sf(crs = st_crs(26910)) # set UTM zone 10 projection

# Publication plot
# shp %>%
#    mutate(HUC8 = as.numeric(HUC8)) %>% 
#    inner_join(cor.huc.max[cor.huc.max$type == 'pekel', ], c("HUC8" = "huc8.img")) %>% 
#    ggplot() + 
#      geom_sf(aes(fill = cor), size = 0.1) +
#      theme_bw(base_size = 10) +
#      theme(plot.margin=grid::unit(c(2, 0, 0, 0), "mm"),  #(T, R, B, L)
#            legend.text = element_text(size = 8),
#            legend.key = element_rect(size = 1, fill = "white", color = "white"),
#            legend.key.size = unit(1.1, 'lines'),
#            legend.spacing.y = unit(1, 'mm')) +  
#      scale_size_identity() +
#      scale_y_continuous(breaks = c(35, 37, 39, 41)) +
#      scale_x_continuous(breaks = c(-122, -120, -118)) +
#      scale_fill_manual(name = "Imagery \ndataset", 
#                          #values= c('orange', 'lightblue'), 
#                          values = c('#F8766D', '#00BFC4'), 
#                          labels = c('DSWE', "JRC")) +
#      scale_fill_gradient(name = "Correlation", limits = c(0,1), low = 'green', high = 'blue') +
#    #  coord_sf(crs = st_crs(26910)) # set UTM zone 10 projection
#      coord_sf(crs = "+init=epsg:4269")  #NAD83 GEO   
# 
# 
# # Save for publication
#  ggsave(file = file.path(path.data, "plots", paste0("jrc_gage_correlations_", cor_type, ".pdf")),
#          width = 3.5,
#          height = 3.8,
#          units = "in",
#          dpi = 600)



```

<br>


####Interpolated plots - within HUC
```{r, echo= FALSE, warning = FALSE, message = FALSE, progress_bar = FALSE, fig.height=7, fig.width=8}

# This can be rewritten such that the pk_ds_sg calls are dropped--the incoming data is already partitioned from that data frame. Leaving it for now because it works.

plot_function  <- function(max_cor) {
  
  huc_i <- max_cor$huc8.img
  site_i <- max_cor$site_no
  huc_name <- paste0("HUC ", huc_i, ", Site ", site_i)

# all imagery data - recorded and interpolated
  pk_ds_sg.all <- subset(pk_ds_sg.int, huc8.img == huc_i & site_no == site_i)

# recorded only
  pk_ds_sg.rec <- pk_ds_sg.all %>% mutate(water = replace(water, which(interp == 'Y'), NA))

# interpolated only
  pk_ds_sg.inp <- pk_ds_sg.all %>% mutate(water = replace(water, which(interp == 'N'), NA))

# combine all and recorded 
  pk_combined <- rbind(pk_ds_sg.all, pk_ds_sg.rec)
 
# change names
# thanks, ttps://stackoverflow.com/questions/3472980/ggplot-how-to-change-facet-labels
  imagery.names <-  list('dswe'= "DSWE", 'pekel' = 'JRC')
  imagery_labeller <- function(variable, value){
    return(imagery.names[value])}
  
  plot.x.max <- max(pk_ds_sg.all$date)
  plot.y.max <- max(pk_ds_sg.all$water)
  labels <-  pk_ds_sg.inp %>% group_by(type) %>% summarize(r = round(mean(cor), 2)) %>% .$r
  plot_label <-  sprintf("rho == %0.2f", labels)

# Original plot
  p <- ggplot(pk_ds_sg.all, aes(date, water, group = type)) +  # all
        geom_line(size = .3, color = 'black') +
        geom_line(data = pk_ds_sg.inp, aes(date, water), color = 'red', size = .8) +
        geom_point(data = pk_ds_sg.all, aes(date, water, color = interp), size = 0.6) +  #all
        scale_color_manual(values = c("black", "red"), labels = c("Measured data", "Estimated data")) +
        theme_bw() +
        theme(plot.margin=unit(c(1, 1, 1, 1),"cm"),
          legend.position =  c(0.96, 1.04), #'top', 
          legend.justification = 'right',
          legend.direction = 'horizontal',
          legend.title = element_blank(),
          legend.background = element_rect(fill = NA)) +
        xlab("Date") +
        ylab('Surface water extent (sq km)')  +
        ggtitle(huc_name) +
    
#legacy code for incorporating water data
#  geom_text(data = max_cor, aes(x = plot.x.min, y = plot.y.min - 10, label = site_no), inherit.aes = FALSE) +
#  scale_y_continuous(sec.axis = sec_axis(~.*300, name = "Stream gage discharge\n")) +  # 4.5
#  geom_area(data = pk_ds_sg.all, aes(date, mean_va/300), color = "blue",  alpha = 0.1)

     
      facet_grid(type ~ ., labeller = imagery_labeller, switch = 'y') +
      theme(strip.background = element_rect(color="black", fill="white", linetype="solid")) +
      annotate("text", 
               label = plot_label,
               parse = T,
               size = 4, 
               x = as.Date(plot.x.max - 300, '%Y-%m-%d'), 
               y = 0.95 * plot.y.max) 
  
  
  print(p)

  ggsave(file = file.path(path.data, "plots", paste0("dswe_jrc_huc", huc_i, ".png")), 
        width = 5,  #7 
        height = 3.5, #5
        dpi = 600)

}

```
<br>



```{r, echo=FALSE}


## *************************
#### plots for publication 
## *************************

plot_function_pub  <- function(max_cor) {
  
  huc_i <- max_cor$huc8.img
  site_i <- max_cor$site_no
  huc_name <- paste0("HUC ", huc_i, ", Gage ", site_i)

# all imagery data - recorded and interpolated
  pk_ds_sg.all <- subset(pk_ds_sg.int, huc8.img == huc_i & site_no == site_i)

# recorded only
  pk_ds_sg.rec <- pk_ds_sg.all %>% mutate(water = replace(water, which(interp == 'Y'), NA))

# interpolated only
  pk_ds_sg.inp <- pk_ds_sg.all %>% mutate(water = replace(water, which(interp == 'N'), NA))

# combine all and recorded 
  pk_combined <- rbind(pk_ds_sg.all, pk_ds_sg.rec)
 
# assign names
  imagery.names <-  list('dswe'= "DSWE", 'pekel' = 'JRC')
  imagery_labeller <- function(variable, value){
    return(imagery.names[value])}
  
  plot.x.max <- max(pk_ds_sg.all$date)
  plot.y.max <- max(pk_ds_sg.all$water)
  labels <-  pk_ds_sg.inp %>% group_by(type) %>% summarize(r = round(mean(cor), 2)) %>% .$r
  plot_label <-  sprintf("rho == %0.2f", labels)


  # Plots for publication

  p <- ggplot(pk_ds_sg.all, aes(date, water, group = type)) +  # all
        geom_line(lwd = 0.3, color = 'black') +
        geom_line(data = pk_ds_sg.inp, aes(date, water), color = 'red', size = 0.3) +
        geom_point(data = pk_ds_sg.all, aes(date, water, color = interp), size = 0.5) +  #all
        scale_color_manual(values = c("black", "red"), labels = c("Measured data", "Estimated data")) +
        theme_bw(base_size = 8.5) +
        guides(colour=guide_legend(override.aes=list(size = 2))) + # set legend point size
        theme(plot.margin=unit(c(0.3, .1, 0, 0.1), "cm"),
          legend.position =  c(0.96, 1.04), #'top', 
          legend.justification = 'right',
          legend.direction = 'horizontal',
          legend.title = element_blank(),
          legend.background = element_rect(fill = NA),
          legend.text = element_text(size = 8),
          legend.key.size = unit(0.7, 'line'),
          legend.key.width = unit(0.9, 'line')) +
        xlab("Date") +
        ylab('Surface water extent (sq km)')  +
        ggtitle(huc_name) +

      facet_grid(type ~ ., labeller = imagery_labeller, switch = 'y') +
      theme(strip.background = element_rect(color = "black", fill = "white", linetype = "solid")) +
      annotate("text", 
               label = plot_label,
               parse = T,
               size = 3, 
               x = as.Date(plot.x.max - 100, '%Y-%m-%d'), #300
               y = 0.95 * plot.y.max) 

    
  print(p)

  ggsave(file = file.path(path.data, "plots", paste0("dswe_jrc_huc_publication", huc_i, ".pdf")), 
        width = 5.2,  #7 
        height = 3.4, #5
        units = "in",
        dpi = 600)
  

}

```


```{r, echo = FALSE, message = FALSE, warning = FALSE, progress_bar = FALSE}

# ---------------------------------------------------------------------- #
#### Generate subset of gage::imagery time-series plots for publication
# ---------------------------------------------------------------------- #

pk_ds_sg.int %>% 
    filter(huc8.sg == huc8.img) %>% 
    #filter(huc8.img == '18040010') %>% 
    filter(huc8.img %in% c('18020004', '18020002', '18040010')) %>% 
    group_by(huc8.img) %>% 
    slice(which.max(cor))  %>% 
    do(df.select = plot_function_pub(.)) # 'do' to pass the groups individually; 'df.select' b/c a name is necessary for some reason

```
<br>



```{r, echo = FALSE, message = FALSE, warning = FALSE, progress_bar = FALSE}

# ---------------------------------------------------------------------- #
#### Generate all gage::imagery time-series plots
# ---------------------------------------------------------------------- #

#huc_acreage <- shp %>% 
#  distinct(HUC8, km2) %>% 
#  mutate(HUC8 = as.numeric(HUC8))

pk_ds_sg.int %>% 
#    left_join(select(huc_acreage, HUC8, km2), c('huc8.img' = 'HUC8')) %>% #
#    select(-geometry) %>% #
    filter(huc8.sg == huc8.img) %>% 
    group_by(huc8.img) %>% 
    slice(which.max(cor))  %>% 
    do(df.select = plot_function(.)) # 'do' to pass the groups individually; 'df.select' b/c a name is necessary for some reason

```
<br>


####Interpolated plots - all HUCs
```{r, echo= FALSE, warning = FALSE, message = FALSE, progress_bar = FALSE, fig.height=7, fig.width=8}

# This can be rewritten such that the pk_ds_sg calls are dropped--the incoming data is already partitioned from that data frame. Leaving it for now because it works.

plot_function_all  <- function(max_cor) {
  
  huc_i <- max_cor$huc8.img
  site_i <- max_cor$site_no
  huc_name <- paste0("HUC ", huc_i, ", Site ", site_i)

# all imagery data  for subset - recorded and interpolated
  pk_ds_sg.all <- subset(pk_ds_sg.int, huc8.img == huc_i & site_no == site_i)

# recorded only
  pk_ds_sg.rec <- pk_ds_sg.all %>% mutate(water = replace(water, which(interp == 'Y'), NA))

# interpolated only
  pk_ds_sg.inp <- pk_ds_sg.all %>% mutate(water = replace(water, which(interp == 'N'), NA))

# combine all and recorded 
  pk_combined <- rbind(pk_ds_sg.all, pk_ds_sg.rec)
 
# change names
# thanks, ttps://stackoverflow.com/questions/3472980/ggplot-how-to-change-facet-labels
  imagery.names <-  list('dswe'= "DSWE", 'pekel' = 'JRC')
  imagery_labeller <- function(variable, value){
    return(imagery.names[value])}
  
  plot.x.max <- max(pk_ds_sg.all$date)
  plot.y.max <- max(pk_ds_sg.all$water)
  labels <-  pk_ds_sg.inp %>% group_by(type) %>% summarize(r = round(mean(cor), 2)) %>% .$r
  plot_label <-  sprintf("rho == %0.2f", labels)

# Plot
  p <- ggplot(pk_ds_sg.all, aes(date, water, group = type)) +  # all
        geom_line(size = .3, color = 'black') +
        geom_line(data = pk_ds_sg.inp, aes(date, water), color = 'red', size = .8) +
        geom_point(data = pk_ds_sg.all, aes(date, water, color = interp), size = 0.6) +  #all
        scale_color_manual(values = c("black", "red"), labels = c("Measured data", "Estimated data")) + 
        theme_bw() +
        theme(plot.margin=unit(c(1, 1, 1, 1),"cm"),
          legend.position =  c(0.96, 1.04), #'top', 
          legend.justification = 'right',
          legend.direction = 'horizontal',
          legend.title = element_blank(),
          legend.background = element_rect(fill = NA)) +
        xlab("Date") +
        ylab('Surface water extent (sq km)')  +
        ggtitle(huc_name) +
    
#legacy code for incorporating water data
#  geom_text(data = max_cor, aes(x = plot.x.min, y = plot.y.min - 10, label = site_no), inherit.aes = FALSE) +
#  scale_y_continuous(sec.axis = sec_axis(~.*300, name = "Stream gage discharge\n")) +  # 4.5
#  geom_area(data = pk_ds_sg.all, aes(date, mean_va/300), color = "blue",  alpha = 0.1)

      facet_grid(type ~ ., labeller = imagery_labeller, switch = 'y') +
      theme(strip.background = element_rect(color="black", fill="white", linetype="solid")) +
      annotate("text", 
               label = plot_label,
               parse = T,
               size = 4, 
               x = as.Date(plot.x.max - 300, '%Y-%m-%d'), 
               y = 0.95 * plot.y.max) 

  print(p)

  ggsave(file = file.path(path.data, "plots", paste0("dswe_jrc_huc", huc_i, "_all_gages.png")), 
        width = 7, 
        height = 5, 
        dpi = 600)

}

```

<br>

```{r, echo = FALSE, message = FALSE, warning = FALSE, progress_bar = FALSE}

# ---------------------------------------------------------------------- #
#### Generate all gage::imagery time-series plots
# ---------------------------------------------------------------------- #

pk_ds_sg.int %>% 
#    left_join(select(huc_acreage, HUC8, km2), c('huc8.img' = 'HUC8')) %>% #
#    select(-geometry) %>% #
#    filter(huc8.sg == huc8.img) %>% 
    group_by(huc8.img) %>% 
    slice(which.max(cor))  %>% 
    do(df.select = plot_function_all(.)) # 'do' to pass the groups individually; 'df.select' b/c a name is necessary for some reason

```
<br>


```{r, echo = FALSE, include = FALSE}
 ggplot(data = pk_ds_sg.int, aes(date, mean_va, group = type)) + 
   scale_fill_manual("Dataset", values = c("blue", "lightblue")) + #, lab#) +
   geom_area(color = 'blue', alpha = 0.1, aes(color = huc8)) + 
   theme_bw() +
   xlab("Date") +
   ylab('Discharge (mean/month)') 
```




```{r, echo = FALSE}

# PARKED 

#Where do the DSWE/Pekel correlations (imagery:gage) differ? 

####Absolute differences between DSWE/gage and Pekel/gage r values

# ---------------------------------------------------------------------- #
#### Spatial plot of Pekel/gage : DSWE/gage correlation differences ####
# ---------------------------------------------------------------------- #
# cor_diffs <- merge(cor_ds_sg_max, cor_pk_sg_max, by = c('huc8', 'parameter_cd'))
# 
# cor_diffs$diff <- abs(cor_diffs$cor.x - cor_diffs$cor.y)
# 
# tmp <- merge(shp, cor_diffs[, c('huc8', 'diff')], by.x = 'HUC8', by.y = 'huc8')
# 
# ggplot(tmp) + 
#      geom_sf(aes(fill = diff)) +
#      theme_bw() +
#      coord_sf(crs = st_crs(26910)) # set UTM zone 10 projection
```



  
```{r, echo = FALSE, warning = FALSE}

#PARKED


#printPlot <- function(df)  {


#
#
# # all imagery data - recorded and interpolated
# pk_ds_sg.all <- subset(pk_ds_sg, huc8 == '18020002' & site_no == 11351945) # & type == 'dswe')
#
# # recorded only
# pk_ds_sg.rec <- pk_ds_sg.all %>% mutate(water = replace(water, which(interp == 'Y'), NA))
#
# # interpolated only
# pk_ds_sg.int <- pk_ds_sg.all %>% mutate(water = replace(water, which(interp == 'N'), NA))
#
# # combine all and recorded
# pk_combined <- rbind(pk_ds_sg.all, pk_ds_sg.rec)
#
# huc.name <- paste0("HUC_", pk_ds_sg.all$huc8[1])
#
#
# # change names
# # thanks, ttps://stackoverflow.com/questions/3472980/ggplot-how-to-change-facet-labels
# imagery.names <-  list('dswe'= "DSWE", 'pekel' = 'JRC/Pekel')
# imagery_labeller <- function(variable,value){
#   return(imagery.names[value])
# }
#
#
# # Plots
# p <- ggplot(pk_ds_sg.all, aes(date, water, group = type)) +
#       geom_line(size = .3, color = 'black') +
#       geom_line(data = pk_ds_sg.int, aes(date, water), color = 'red', size = .8) +
#       geom_point(data = pk_ds_sg.all, aes(date, water, color = interp), size = 0.6) +
# #  geom_line(aes(linetype = type)) +
# #  scale_linetype_manual(values = c(1,1), labels = c("Measured data", "Estimated data")) +
#
#       scale_color_manual(values = c("black", "red"), labels = c("Measured data", "Estimated data")) +
#  #   guides(linetype = guide_legend(override.aes = list(alpha=1))) +
#       theme_bw() +
#       theme(plot.margin=unit(c(1, 1, 1, 1),"cm"),
#         legend.position =  c(0.95, 1.062), #'top',
#         legend.justification = 'right',
#         legend.direction = 'horizontal',
#         legend.title = element_blank()) +
#       xlab("Date") +
#       ylab('Monthly extent (sq km)')  +
#       ggtitle(huc.name) +
#     scale_y_continuous(sec.axis = sec_axis(~.*300, name = "Stream gage discharge\n")) +  # 4.5
#   geom_area(data = pk_ds_sg.all, aes(date, mean_va/300), color = "blue",  alpha = 0.1)
#
#
# p + facet_grid(type ~ ., labeller=imagery_labeller, switch = 'y') +
#    theme(strip.background = element_rect(color="black", fill="white", linetype="solid"))
# +
#
#   # second axis! https://rpubs.com/MarkusLoew/226759
#
#   scale_y_continuous(sec.axis = sec_axis(~.*4.5, name = "Stream gage discharge\n")) +
#   geom_area(data = pk_ds_sg.all, aes(date, mean_va/4.5), color = "blue",  alpha = 0.1)

```
